# config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Ecommerce",
    "version": "1.0.0",
    "description": "Documentación de la API para la tienda Ecommerce"
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Servidor de Desarrollo"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "ID único del producto",
            "example": 1
          },
          "nombre": {
            "type": "string",
            "description": "Nombre del producto",
            "example": "Smartphone XYZ"
          },
          "descripcion": {
            "type": "string",
            "description": "Descripción del producto",
            "example": "Smartphone de última generación con múltiples funcionalidades."
          },
          "precio": {
            "type": "number",
            "format": "float",
            "description": "Precio del producto",
            "example": 299.99
          },
          "cantidadEnStock": {
            "type": "integer",
            "description": "Cantidad disponible en stock",
            "example": 50
          },
          "categoriaId": {
            "type": "integer",
            "description": "ID de la categoría a la que pertenece el producto",
            "example": 1
          },
          "imagenes": {
            "type": "array",
            "items": {
              "type": "string",
              "description": "Nombre de archivo de las imágenes del producto",
              "example": "smartphone1.jpg"
            }
          },
          "featured": {
            "type": "boolean",
            "description": "Indica si el producto es destacado",
            "example": true
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/products": {
      "get": {
        "summary": "Obtener todos los productos con opcional filtro por categoría o destacados",
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "categoria",
            "schema": {
              "type": "string"
            },
            "description": "Nombre de la categoría para filtrar productos"
          },
          {
            "in": "query",
            "name": "featured",
            "schema": {
              "type": "boolean"
            },
            "description": "Filtrar productos destacados"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de productos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error al obtener los productos"
          }
        }
      },
      "post": {
        "summary": "Crear un nuevo producto (Solo Administradores)",
        "tags": [
          "Products"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "nombre",
                  "precio",
                  "cantidadEnStock",
                  "categoriaId"
                ],
                "properties": {
                  "nombre": {
                    "type": "string",
                    "description": "Nombre del producto"
                  },
                  "descripcion": {
                    "type": "string",
                    "description": "Descripción del producto"
                  },
                  "precio": {
                    "type": "number",
                    "format": "float",
                    "description": "Precio del producto"
                  },
                  "cantidadEnStock": {
                    "type": "integer",
                    "description": "Cantidad en stock"
                  },
                  "categoriaId": {
                    "type": "integer",
                    "description": "ID de la categoría"
                  },
                  "imagenes": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "Imágenes del producto"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Producto creado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Error de validación"
          },
          "403": {
            "description": "Acceso denegado"
          },
          "500": {
            "description": "Error al crear el producto"
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "summary": "Obtener detalles de un producto por ID",
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID del producto"
          }
        ],
        "responses": {
          "200": {
            "description": "Detalles del producto",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "404": {
            "description": "Producto no encontrado"
          },
          "500": {
            "description": "Error al obtener el producto"
          }
        }
      },
      "put": {
        "summary": "Actualizar un producto existente (Solo Administradores)",
        "tags": [
          "Products"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID del producto a actualizar"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombre": {
                    "type": "string",
                    "description": "Nombre del producto"
                  },
                  "descripcion": {
                    "type": "string",
                    "description": "Descripción del producto"
                  },
                  "precio": {
                    "type": "number",
                    "format": "float",
                    "description": "Precio del producto"
                  },
                  "cantidadEnStock": {
                    "type": "integer",
                    "description": "Cantidad en stock"
                  },
                  "categoriaId": {
                    "type": "integer",
                    "description": "ID de la categoría"
                  },
                  "imagenes": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "Nuevas imágenes del producto"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Producto actualizado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Error de validación"
          },
          "403": {
            "description": "Acceso denegado"
          },
          "404": {
            "description": "Producto no encontrado"
          },
          "500": {
            "description": "Error al actualizar el producto"
          }
        }
      },
      "delete": {
        "summary": "Eliminar un producto (Solo Administradores)",
        "tags": [
          "Products"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID del producto a eliminar"
          }
        ],
        "responses": {
          "200": {
            "description": "Producto eliminado exitosamente"
          },
          "403": {
            "description": "Acceso denegado"
          },
          "404": {
            "description": "Producto no encontrado"
          },
          "500": {
            "description": "Error al eliminar el producto"
          }
        }
      }
    },
...
  },
...
}